+++ EntityManager.h
@@ -11,7 +11,7 @@ class EntityManager {
 public:
     EntityManager(); // Constructor
     void init(); // Initialize the entities
-    virtual void update(Arduboy2 &arduboy, int xDelta); // Update all entities
+    virtual void update(Arduboy2 &arduboy, float xDelta); // Update all entities
     void draw(Arduboy2 &arduboy); // Draw all entities
 
     // Add a new entity to the manager
